SharePoint_Report-Storage_Sharing_Members.ps1
Changed from saving to "C:\Tmp" to "C:\Temp".
Added a check to confirm the folder "C:\Temp" exists. 
Updated "Storage Used" column to GB, which is now calculated by /1024, as previous was reported wrong. 
Added a column to report on members of the site (Including members added via a group).